fix: change methods and operators as part of OMNeT++ 6 compatibility

+++ playground/DAF/AE/AESInt.cc
@@ -86,7 +86,7 @@ void AESInt::initialize()
 
 
     //MAX PDU Size
-    maxSize = par("maxPDUsize").longValue();
+    maxSize = par("maxPDUsize").intValue();
 
     send = 0;
     sendSize = 0;
