+++ src/DAF/AEManagement/AEMgmtBase.cc
@@ -44,7 +44,7 @@ void AEMgmtBase::initMyAddress() {
 }
 
 long AEMgmtBase::getNewInvokeId() {
-    long newinvoke = getParentModule()->getSubmodule(MOD_CDAP)->par(PAR_CURINVOKEID).longValue() + 1;
+    long newinvoke = getParentModule()->getSubmodule(MOD_CDAP)->par(PAR_CURINVOKEID).intValue() + 1;
     getParentModule()->getSubmodule(MOD_CDAP)->par(PAR_CURINVOKEID) = newinvoke;
     return newinvoke;
 }
